# react-app/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AspNotes API",
    "version": "v1"
  },
  "paths": {
...
    "/api/books/getList": {
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Retrieves a list of books based on the provided section.",
        "requestBody": {
          "description": "The request containing the section to filter books.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetBooksListRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetBooksListRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetBooksListRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the list of books.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookItemResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If the section name is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/books/autocomplete": {
      "post": {
        "tags": [
          "Books"
        ],
        "summary": "Provides book name suggestions based on the input term and optional section.",
        "requestBody": {
          "description": "The autocomplete request containing the search term and optional section.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutocompleteBooksRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AutocompleteBooksRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AutocompleteBooksRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the list of suggested book names.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "If the section name is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/api/sources/add": {
      "post": {
        "tags": [
          "Sources"
        ],
        "summary": "Adds a new source to a note.",
        "requestBody": {
          "description": "The request containing the note ID and source URL to add.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteSourceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteSourceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the updated list of sources for the note.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "400": {
            "description": "If the note is not found or the request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/sources/update": {
      "post": {
        "tags": [
          "Sources"
        ],
        "summary": "Updates an existing source of a note.",
        "requestBody": {
          "description": "The request containing the note ID, source ID, and new source details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteSourceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteSourceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNoteSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the updated list of sources for the note.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "400": {
            "description": "If the note or source is not found or the request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/sources/remove": {
      "post": {
        "tags": [
          "Sources"
        ],
        "summary": "Removes a source from a note.",
        "requestBody": {
          "description": "The request containing the note ID and source ID to remove.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveNoteSourceRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveNoteSourceRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveNoteSourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the updated list of sources for the note.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "400": {
            "description": "If the note or source is not found or the request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/sources/reorder": {
      "post": {
        "tags": [
          "Sources"
        ],
        "summary": "Reorders the sources of a note.",
        "requestBody": {
          "description": "The request containing the note ID and a list of source IDs in the new order.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderNoteSourcesRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderNoteSourcesRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderNoteSourcesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns the reordered list of sources for the note.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            }
          },
          "400": {
            "description": "If the note is not found, a source is not found, or the request is invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNoteSourceRequest": {
        "required": [
          "link",
          "noteId"
        ],
        "type": "object",
        "properties": {
          "noteId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "link": {
            "minLength": 1,
            "type": "string",
            "format": "uri"
          }
        },
        "additionalProperties": false
      },
...
      "AutocompleteBooksRequest": {
        "required": [
          "searchTerm"
        ],
        "type": "object",
        "properties": {
          "searchTerm": {
            "minLength": 2,
            "type": "string"
          },
          "section": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "BookItemResponse": {
        "required": [
          "count",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "count": {
            "type": "integer",
            "format": "int64"
          }
        },
        "additionalProperties": false
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "showNotification": {
            "type": "boolean"
          },
          "field": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "GetBooksListRequest": {
        "type": "object",
        "properties": {
          "section": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RemoveNoteSourceRequest": {
        "required": [
          "noteId",
          "sourceId"
        ],
        "type": "object",
        "properties": {
          "noteId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "sourceId": {
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ReorderNoteSourcesRequest": {
        "required": [
          "noteId",
          "sourceIds"
        ],
        "type": "object",
        "properties": {
          "noteId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "sourceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "SourceItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "link": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "showImage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "SourcesResponse": {
        "required": [
          "sources"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceItemResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateNoteSourceRequest": {
        "required": [
          "link",
          "noteId",
          "sourceId"
        ],
        "type": "object",
        "properties": {
          "noteId": {
            "minimum": 1,
            "type": "integer",
            "format": "int64"
          },
          "sourceId": {
            "minLength": 1,
            "pattern": "\\S",
            "type": "string"
          },
          "link": {
            "minLength": 1,
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "showImage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme",
        "scheme": "bearer"
      }
    }
  }
}